Move login page and adjust navbars
